sugar tests fix, bugfix envs
