+++ .gitignore
@@ -40,6 +40,7 @@ dmypy.json
 .tox/
 .nox/
 nosetests.xml
+allure-results/
 
 # Airflow runtime artifacts (in case someone runs this near an AIRFLOW_HOME)
 airflow.cfg
