+++ pytest.ini
@@ -1,0 +1,13 @@
+# Pytest configuration. Takes precedence over any [tool.pytest.ini_options] in
+# pyproject.toml, so the project's pytest settings live here (single source).
+[pytest]
+testpaths = tests
+pythonpath = src
+filterwarnings =
+    ignore:cannot collect test class 'TestRunResult':pytest.PytestCollectionWarning
+
+# Allure reports are OPT-IN, kept out of addopts so the bare-pytest CI job (no
+# allure-pytest installed) stays green. Set ALLURE_DIR to auto-generate them on
+# every run -- the injection is handled in tests/conftest.py and needs the plugin:
+#     pip install "airflow-pytest-operator[pytest-allure]"
+#     ALLURE_DIR=allure-results pytest
