+++ requirements/dev.txt
@@ -1,16 +1,3 @@
-#
-# Hash-pinned dev/test toolchain for CI. Mirrors the [dev] extra in
-# pyproject.toml (the single source of truth) and is installed with:
-#     pip install --require-hashes -r requirements/dev.txt
-# so OpenSSF Scorecard's Pinned-Dependencies check sees every CI dependency
-# pinned by hash. Universal lock: markers cover py3.10-3.13 and all platforms,
-# so one file serves the whole CI matrix.
-#
-# Regenerate after editing the [dev] extra in pyproject.toml:
-#     uv pip compile pyproject.toml --extra dev --universal --generate-hashes \
-#       --python-version 3.10 --no-annotate --no-header -o requirements/dev.txt
-# Dependabot (pip ecosystem) also bumps these hashes via PRs.
-#
 ast-serialize==0.5.0 \
     --hash=sha256:061ee58bdb52341c8201a6df41182a977736bae3b7ded87ca7176ca25a8a47ab \
     --hash=sha256:0668aa9459cfa8c9c49ddd2163ebcf43088ba045ef7492af6fe22e0098303101 \
@@ -146,6 +133,9 @@ defusedxml==0.7.1 \
 exceptiongroup==1.3.1 ; python_full_version < '3.11' \
     --hash=sha256:8b412432c6055b0b7d14c310000ae93352ed6754f70fa8f7c34141f91c4e3219 \
     --hash=sha256:a7a39a3bd276781e98394987d3a5701d0c4edffb633bb7a5144577f82c773598
+execnet==2.1.2 \
+    --hash=sha256:63d83bfdd9a23e35b9c6a3261412324f964c2ec8dcd8d3c6916ee9373e0befcd \
+    --hash=sha256:67fba928dd5a544b783f6056f449e5e3931a5c378b128bc18501f7ea79e296ec
 iniconfig==2.3.0 \
     --hash=sha256:c76315c77db068650d49c5b56314774a7804df16fee4402c1f19d6d15d8c4730 \
     --hash=sha256:f631c04d2c48c52b84d0d0549c99ff3859c98df65b3101406327ecc7d53fbf12
@@ -312,6 +302,9 @@ pytest-json-report==1.5.0 \
 pytest-metadata==3.1.1 \
     --hash=sha256:c8e0844db684ee1c798cfa38908d20d67d0463ecb6137c72e91f418558dd5f4b \
     --hash=sha256:d2a29b0355fbc03f168aa96d41ff88b1a3b44a3b02acbe491801c98a048017c8
+pytest-xdist==3.8.0 \
+    --hash=sha256:202ca578cfeb7370784a8c33d6d05bc6e13b4f25b5053c30a152269fd10f0b88 \
+    --hash=sha256:7e578125ec9bc6050861aa93f2d59f1d8d085595d6551c2c90b6f4fad8d3a9f1
 python-dotenv==1.2.2 \
     --hash=sha256:1d8214789a24de455a8b8bd8ae6fe3c6b69a5e3d64aa8a8e5d68e694bbcb285a \
     --hash=sha256:2c371a91fbd7ba082c2c1dc1f8bf89ca22564a087c2c287cd9b662adde799cf3
